Add piece drop
Add piece drop function. This is now functional. It still needs some
restrictions placed on it, some guards and assertions need to be added
otherwise it can behave badly due to the bit fiddling and int usage
going on. Good start though. Basis for this function is almost directly
from Thierry and Scherrer.