Env skeleton and piece loading
Added a skeleton for the first environment and wrote the function to
load pieces from a file.

This uses the same file format as Thiery and Scherrer. Tetris shapes
need to give number of orientations, height, and width.

Next to do is to set up a basic game, flesh out the step function etc.
This setup should work for small board games and variety of piece
sets. Next issue is the action space, but I have a plan for that.