Testing step function
Test step function. Currently the performance is that 100,000 pieces can
be dropped every 10 seconds. In comparison the Thierry and Scherrer C
version can complete 100,000 lines every second. Considering it takes on
average 2.5 piece drops to clear a line, my version is currently running
25x slower. This is probably an acceptable level of speed to perform at,
although it would be good to look at converting some of the files to
Cython to see if any speed improvements can be made. Next job is to also
shape the env to an actual openai gym environment.