Added observation space and action space
Added observation space and action space environment variables for
compatability with OpenAI gym. Ready to try and test with a stable
baselines algorithm implmentation. Next to try is probably to abstract
the actual Tetris class and use it to implement the Melax reduced
environment for initial training. After that feature extraction for
linear heuristics would be good, potentially followed by some Cython
optimisation of the core functions. The way the current state is
returned and all of the board functions would likely be speeded up by
some injection of Cython into the mix. Sprinklings of cdef.