Cythonization of engine functions
This commit cythonizes the main processes of the board and provides a
10X speedup for piece drop and other board related methods! I'm now a
massive fan of Cython.

It still takes 6s to run 100,000 steps of the action, but this is
because the environment file itself is not yet cythonized. Once this is
actioned the speed improvements should further increase.

The board is now represented as a C array of unsigned ints instead of a
numpy array.

One issue I ran into was exporting static variables, in particular the
brick masks. The resolution of this was to create a `.pxd` header file
to declare the variables I wanted to import from `brick_masks.pyx`