Cythonize Tetris
Initial cythonization of the Tetris environment. Now the main functions
of the board and pieces are cythonized, the next step is to finish up
the class.

Multiple inheritance can be used to ensure the class can still be used
as a Gym environment. Important to go a step at a time with continuous
building and testing here.