Speed improvements
Increased speed by 3x again (from 6s to 2s for 100,000 piece drops and
resets). Changes made to do this:
- cdef clamp function. This clamps the value while avoiding python funcs
- Python random. Got rid of the numpy random choice in favour of
Python's built in random for a new_piece fucntion that selects a new
piece much more quickly for both the step and reset functions.

Remaining overhead is apparent in the numpy concatenate that returns the
state, it may be possible to eliminate that? Overhead in actual piece
drop is minimal.