Convert to binary state array
Developed an initial method that converts the state array from the
integer representation to a binary representation.

There might be a better way of going back to the board class, rethinking
and rewriting it there. If it's natively a numpy array then it should be
quicker to work with?

Also the observation space remains a problem as a result. How should the
piece be shown? I don't fully understnad gym.spaces and how to
represent the observation space.