Feature set initial commit
Starting to add functions to enable the Dellacherie feature set
computation. Once this is finished (probably soon) I'll be able to
implement a linear Tetris environment where the 'action' is the
selection of the feature weights followed by game runouts, that could be
interesting.

As it is it will also provide me a good baseline for comparing the
performance of other policies.