Redo state observation and add flat vers
The state observation has been redone as a single 2D binary array that
captures the detail of the current piece.

The top left corner of the board has the piece shaape placed there. The
state representation then returns the 2D array that results, or athe
flattened 1D array if the flat environment is used.