Update observation spaces
Each environment extension now has a `get_observation_space` function to
indicate its observation space.

The heuristic environments currently do not meet the MDP property of
giving all required information, as they only present the heuristic
rating of the current state without giving the state of the board or the
current piece.

Thinking about how one way to get around that would be to present an
array of the heuristic values of all potential future states, indexed by
the actions that lead to them. This would be computationally costly, but
is one way to present the representation as a gym env MDP.